+++ dokumentaatio/changelog.md
@@ -7,3 +7,8 @@
 - Lisätty YarnService-luokka
 - Käyttäjä pystyy lisätä langan varastoon
 - Testaus aloitettu; TodoService-luokka toimii langan lisäämisen osalta
+
+## Viikko 4
+
+- Käyttäjä näkee varastoon lisäämänsä langat
+- Käyttäjä voi siirtyä etusivulta joko lisäämään langan tai katsoa kaikki varastossa olevat langat